# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/categories": {
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Category name already exists"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a new category",
        "tags": [
          "categories"
        ]
      },
      "get": {
        "operationId": "CategoriesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get all categories",
        "tags": [
          "categories"
        ]
      }
    },
    "/categories/{id}": {
      "get": {
        "operationId": "CategoriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a category by ID",
        "tags": [
          "categories"
        ]
      },
      "patch": {
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          },
          "409": {
            "description": "Category name already exists"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a category",
        "tags": [
          "categories"
        ]
      },
      "delete": {
        "operationId": "CategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a category",
        "tags": [
          "categories"
        ]
      }
    },
...
  },
  "info": {
    "title": "My API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Operation completed successfully",
            "description": "Response message"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Electronics",
            "description": "Category name",
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "example": "Electronic devices and accessories",
            "description": "Category description"
          }
        },
        "required": [
          "name"
        ]
      },
      "CategoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Category unique identifier"
          },
          "name": {
            "type": "string",
            "example": "Electronics",
            "description": "Category name",
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "example": "Electronic devices and accessories",
            "description": "Category description"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Category creation date"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Category last update date"
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Electronics",
            "description": "Category name",
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "example": "Electronic devices and accessories",
            "description": "Category description"
          }
        }
      },
...
    }
  }
}